requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-Admin==1.6.1
flask-cors==5.0.1
Flask-SocketIO==5.5.1
Flask-Turnstile==0.1.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
reportlab==4.4.0
requests==2.32.3
schedule==1.2.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
sqlmodel==0.0.24
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
XlsxWriter==3.2.3
python-dotenv
pycryptodome
openai